+++ src/commands/rest/index.js
@@ -1,0 +1,8 @@
+import myOs from './os.js';
+import hash from './hash.js';
+import { options as help }  from './help.js';
+import exit from './exit.js';
+
+export default {
+  myOs, hash, help, exit,
+};
